Updated argument and variable parsing to use double, updated symbol table
feat: Changed variable assignment and literal creation to parse doubles
refactor: symboltable to have imports in header file
chore: Updated readme to show progress

+++ Project 2/src/operand.cpp
@@ -24,7 +24,7 @@ using namespace std;
 
 Expression* Operand::parse(stringstream& in) {
     char paren;
-    int value;
+    double value;
 
     in >> ws;
     if (isdigit(in.peek())) {
@@ -32,11 +32,13 @@ Expression* Operand::parse(stringstream& in) {
         Expression* literal = new Literal(value);
         return literal;
     }
+
     if (in.peek() == '(') {
         in >> paren;
         return SubExpression::parse(in);
     }
     else
         return new Variable(parseName(in));
+        
     return 0;
 }
